+++ utils/extension-settings/types.ts
@@ -1,0 +1,11 @@
+import { TSchema, Static } from "typebox";
+
+export type ExtensionSettingDefinition<TSettingSchema extends TSchema> = {
+  schema: TSettingSchema;
+  default: () => Static<TSettingSchema>;
+};
+
+export type ExtensionSettingsSchemaMap = Record<
+  string,
+  ExtensionSettingDefinition<TSchema>
+>;